+++ wallet_app/packages/wallet_core/lib/src/models/instruction.dart
@@ -19,7 +19,7 @@ sealed class WalletInstructionError with _$WalletInstructionError {
     required bool isFinalRound,
   }) = WalletInstructionError_IncorrectPin;
   const factory WalletInstructionError.timeout({
-    required int timeoutMillis,
+    required BigInt timeoutMillis,
   }) = WalletInstructionError_Timeout;
   const factory WalletInstructionError.blocked() = WalletInstructionError_Blocked;
 }
